Initial commit
In this commit a basic GoogleSQL mode has been created. This mode
includes syntax highlighting for all GoogleSQL keywords and built-in
functions, and indentation that should cover the full query
syntax (though this has not been fully tested).

Indentation still needs to be added to DML and DDL statements and
constructs, as well as the procedural language.

The code can also use some cleanup and
refactoring. `googlesql-beginning-of-construct' probably does not need
to have a stack, but could be a FSM. The `googlesql-inside-jinja'
function needs to be fixed, and Jinja indentation rules need to be
added. Syntax highlighting should also follow different rules when
inside a jinja block.